fix review bug

+++ common/zero_copy_sink.go
@@ -162,7 +162,7 @@ func (self *ZeroCopySink) WriteAddrList(addrs []Address) {
 }
 
 func (self *ZeroCopySink) WriteVarBytesArray(data [][]byte) {
-	self.WriteUint64(uint64(len(data)))
+	self.WriteUint32(uint32(len(data)))
 	for _, b := range data {
 		self.WriteVarBytes(b)
 	}
